+++ src/pages/register.tsx
@@ -1,24 +1,24 @@
 import { FormInput } from '@smartive-education/pizza-hawaii';
 import Head from 'next/head';
 import Link from 'next/link';
+import LoginLayout from '../components/layoutComponents/LoginLayout';
 
 const RegisterPage = () => {
 
 	return (
-		<>
-			<Head>
-				<title>Register to Mumble</title>
-				<link rel="icon" href="/favicon.ico" />
-			</Head>
-
+		<LoginLayout>
+			<>
+				<Head>
+					<title>Register to Mumble</title>
+					<link rel="icon" href="/favicon.ico" />
+				</Head>
+			</>
 			<main>
-                <h1>Register now</h1>
- 
-               {/* TODO: Form here\ */}
-               <br/ >
-                <Link href='/'>Back to startpage</Link>
+				<h1>Register now</h1>
+				<br />
+				<Link href="/">Back to startpage</Link>
 			</main>
-		</>
+		</LoginLayout>
 	);
 }
